Remove JupyterLite dependencies
