+++ package.json
@@ -20,8 +20,7 @@
     "main": "lib/index.js",
     "types": "lib/index.d.ts",
     "scripts": {
-        "build": "tsc && npm run build:worker",
-        "build:worker": "rspack --config worker.rspack.config.js --mode=development",
+        "build": "tsc",
         "eslint": "npm run eslint:check -- --fix",
         "eslint:check": "eslint . --cache --ext .ts,.tsx",
         "lint": "npm run prettier && npm run eslint",
@@ -32,7 +31,6 @@
         "prettier:check": "prettier --list-different \"**/*{.ts,.tsx,.js,.jsx,.css,.json,.md,.yml}\""
     },
     "dependencies": {
-        "@jupyterlite/contents": "^0.5.1",
         "@lumino/coreutils": "^2.2.0",
         "@lumino/disposable": "^2.1.3",
         "@lumino/signaling": "^2.1.3",
