+++ src/index.ts
@@ -1,4 +1,6 @@
 export { Aliases } from './aliases';
+export { BaseShell } from './base_shell';
+export { BaseShellWorker } from './base_shell_worker';
 export { IOutputCallback, IEnableBufferedStdinCallback, IStdinCallback } from './callback';
 export { CommandRegistry } from './command_registry';
 export { Context } from './context';
